cannot stat /etc/netplan/01-netcfg.yaml
On my system the file /etc/netplan/01-netcfg.yaml is not available after sudo apt install netplan.io. 
We should check if the file exists befor making a backup, otherwise the script exists without a real reason.

+++ static/static_ip.sh
@@ -20,7 +20,7 @@ clear
 msg_box "Copying old netplan.io config file file to:
 
 /tmp/01-netcfg.yaml_backup"
-check_command cp -v /etc/netplan/01-netcfg.yaml /tmp/01-netcfg.yaml_backup
+[[ -f /etc/netplan/01-netcfg.yaml ]] && check_command cp -v /etc/netplan/01-netcfg.yaml /tmp/01-netcfg.yaml_backup
 
 echo
 while true
